+++ Test/MCSTest.cpp
@@ -415,8 +415,8 @@ namespace Test
 
 	void MCSTest::Parallel(IStreamCipher* Cipher)
 	{
-		const size_t MINSMP = Cipher->ParallelBlockSize();
-		const size_t MAXSMP = Cipher->ParallelBlockSize() * 4;
+		const uint MINSMP = static_cast<uint>(Cipher->ParallelBlockSize());
+		const uint MAXSMP = static_cast<uint>(Cipher->ParallelBlockSize()) * 4;
 		Cipher::SymmetricKeySize ks = Cipher->LegalKeySizes()[0];
 		std::vector<byte> cpt1;
 		std::vector<byte> cpt2;
